# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Voice Audit API",
    "description": "\nAI Voice Audit API provides a robust set of endpoints for automating the quality assurance process in call centers using advanced AI models.\n\n## Key Features\n- **Asynchronous Audio Processing**: Upload call recordings for transcription and AI-powered evaluation.\n- **Dynamic LLM Selection**: Configure specific AI models per Line of Business (LOB).\n- **Multi-Tenancy**: Secure data isolation and tenant-specific configurations.\n- **Advanced Analytics**: Real-time dashboards and exportable reports.\n\n## Authentication\nThis API uses **JWT tokens** for secure authentication. \n1. Obtain tokens via the `/api/auth/login` endpoint.\n2. Include the access token in the `Authorization` header for all protected endpoints:\n   `Authorization: Bearer <your_access_token>`\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/lobs/": {
      "get": {
        "tags": [
          "lobs"
        ],
        "summary": "List All LOBs",
        "description": "Retrieve all Lines of Business for the current tenant.\n\n## Example curl:\n```bash\ncurl -H \"Authorization: Bearer <access_token>\" \"http://localhost:5000/api/lobs/\"\n```",
        "operationId": "get_lobs_api_lobs__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LOBResponse"
                  },
                  "type": "array",
                  "title": "Response Get Lobs Api Lobs  Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Insufficient permissions to access this resource"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "lobs"
        ],
        "summary": "Create New LOB",
        "description": "Create a new Line of Business with specific evaluation criteria and system prompts.\n\n## Example curl:\n```bash\ncurl -X POST \"http://localhost:5000/api/lobs/\"          -H \"Authorization: Bearer <token>\"          -H \"Content-Type: application/json\"          -d '{\"name\": \"Support\", \"system_prompt\": \"Helpful...\", \"criteria_json\": {}}'\n```",
        "operationId": "create_lob_api_lobs__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LOBCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Lob Api Lobs  Post"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing required field: email"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Insufficient permissions to access this resource"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/lobs/active": {
      "get": {
        "tags": [
          "lobs"
        ],
        "summary": "List Active LOBs",
        "description": "Retrieve only the active Lines of Business for the current tenant.",
        "operationId": "get_active_lobs_api_lobs_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LOBResponse"
                  },
                  "type": "array",
                  "title": "Response Get Active Lobs Api Lobs Active Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/lobs/{lob_id}/activate": {
      "put": {
        "tags": [
          "lobs"
        ],
        "summary": "Activate LOB",
        "description": "Set a specific LOB as active, deactivating others if necessary (based on tenant policy).",
        "operationId": "activate_lob_api_lobs__lob_id__activate_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lob_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lob Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lobs/{lob_id}": {
      "put": {
        "tags": [
          "lobs"
        ],
        "summary": "Update LOB",
        "description": "Update the name, prompt, or criteria of an existing LOB.",
        "operationId": "update_lob_api_lobs__lob_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lob_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lob Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LOBUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "lobs"
        ],
        "summary": "Delete LOB",
        "description": "Permanently delete a Line of Business.",
        "operationId": "delete_lob_api_lobs__lob_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lob_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lob Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lobs/{lob_id}/llm-config": {
      "get": {
        "tags": [
          "lobs"
        ],
        "summary": "Get LLM Configuration",
        "description": "Retrieve the AI provider configuration for a specific LOB.\n\n**Security Note:** The API key is masked (`\u2022\u2022\u2022\u2022\u2022\u2022\u2022\u2022`) in the response.",
        "operationId": "get_llm_config_api_lobs__lob_id__llm_config_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lob_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lob Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "lobs"
        ],
        "summary": "Save LLM Configuration",
        "description": "Store the AI provider settings (provider, model, API key) for a LOB.\n\nThe API key is encrypted at rest using AES-256.",
        "operationId": "save_llm_config_api_lobs__lob_id__llm_config_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lob_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lob Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lobs/{lob_id}/test-llm": {
      "post": {
        "tags": [
          "lobs"
        ],
        "summary": "Test LLM Connection",
        "description": "Validate that the provided (or stored) LLM configuration can successfully \nreach the AI provider and perform a simple handshake/chat.",
        "operationId": "test_llm_connection_api_lobs__lob_id__test_llm_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lob_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lob Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LLMTestRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "Provider name",
            "examples": [
              "ollama"
            ]
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model name",
            "examples": [
              "llama3.1:8b"
            ]
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "API Key for testing"
          },
          "api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base",
            "description": "Custom API base URL"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model"
        ],
        "title": "LLMTestRequest",
        "description": "Schema for testing an LLM connection."
      },
      "LOBCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the Line of Business",
            "examples": [
              "Customer Support",
              "Sales"
            ]
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id",
            "description": "The tenant ID this LOB belongs to",
            "examples": [
              1
            ]
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt",
            "description": "The system prompt for AI evaluation",
            "examples": [
              "You are a helpful assistant..."
            ]
          },
          "criteria_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Criteria Json",
            "description": "JSON object containing evaluation criteria",
            "examples": [
              {
                "greeting": {
                  "points": 10,
                  "question": "Did the agent greet?"
                }
              }
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether this LOB is currently active",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "system_prompt",
          "criteria_json"
        ],
        "title": "LOBCreate"
      },
      "LOBResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the Line of Business",
            "examples": [
              "Customer Support",
              "Sales"
            ]
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id",
            "description": "The tenant ID this LOB belongs to",
            "examples": [
              1
            ]
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt",
            "description": "The system prompt for AI evaluation",
            "examples": [
              "You are a helpful assistant..."
            ]
          },
          "criteria_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Criteria Json",
            "description": "JSON object containing evaluation criteria",
            "examples": [
              {
                "greeting": {
                  "points": 10,
                  "question": "Did the agent greet?"
                }
              }
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether this LOB is currently active",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Unique LOB identifier",
            "examples": [
              1
            ]
          },
          "is_builtin": {
            "type": "boolean",
            "title": "Is Builtin",
            "description": "Whether this is a system-provided LOB"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "name",
          "system_prompt",
          "criteria_json",
          "id",
          "is_builtin",
          "created_at"
        ],
        "title": "LOBResponse",
        "example": {
          "created_at": "2024-01-01T12:00:00",
          "criteria_json": {
            "greeting": {
              "mandatory": true,
              "points": 5,
              "question": "Did the agent say hello?"
            }
          },
          "id": 1,
          "is_active": true,
          "is_builtin": false,
          "name": "General Support",
          "system_prompt": "Evaluate call center etiquette...",
          "tenant_id": 1
        }
      },
      "LOBUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "New name for the LOB"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt",
            "description": "New system prompt"
          },
          "criteria_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteria Json",
            "description": "Updated criteria JSON"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "New active status"
          }
        },
        "type": "object",
        "title": "LOBUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  },
...
}